Revert "Typescript Setup"
