--- tsconfig.json
@@ -1,14 +1,0 @@
-{
-    "compilerOptions": {
-        "module": "ES2022",
-        "target": "ES2022",
-        "moduleResolution": "Node",
-        "allowSyntheticDefaultImports": true,
-        "esModuleInterop": true
-        /* ... your props ... */
-    },
-    "ts-node": {
-        "esm": true,
-        "experimentalSpecifierResolution": "node",
-    }
-  }
